H2: Add minimal support for trailers
This introduces minimal support for H2 trailers, we no longer
fail a request that contains trailers, but for now we only read
the trailers for HPAC consistency, then we drop them. The workspace of
the struct req associated with the H2 session is used to read the trailers
to avoid any concurent use of the stream req workspace.